+++ tests/PhpSpreadsheetTests/Shared/StringHelperInvalidCharTest.php
@@ -40,5 +40,6 @@ class StringHelperInvalidCharTest extends TestCase
                 $sheet->getCell("A$row")->getValue()
             );
         }
+        $spreadsheet->disconnectWorksheets();
     }
 }