+++ tests/PhpSpreadsheetTests/Style/ConditionalFormatting/Wizard/WizardFactoryTest.php
@@ -67,6 +67,7 @@ class WizardFactoryTest extends TestCase
             $wizard = Wizard::fromConditional($conditional);
             self::assertEquals($expectedWizads[$index], get_class($wizard));
         }
+        $spreadsheet->disconnectWorksheets();
     }
 
     public static function conditionalProvider(): array